Equipment Management
Added an Equipment Script that handles Equipped items by taking the equipmentSlot variable from items and asigning them to the right part of the body.
Equipping items takes them out of the inventory and into the Equipment and back.
So far equipping items can only be done via script  but should later be hooked up to a equipment dialog ui.

also hooked up the chest to the new inventory dialog and removed the old inventory scripts and scenes.
Furthermore made some adjustments to the mouse capture.